add filter to table header
