+++ frontend/packages/core/src/Table/index.tsx
@@ -1,6 +1,7 @@
+import TableCell from "./components/TableCell";
 import { AccordionRow } from "./accordion-table";
 import { MetadataTable } from "./metadata-table";
-import { Table, TableCell, TableRow, TableRowAction, TableRowActions } from "./table";
+import { Table, TableRow, TableRowAction, TableRowActions } from "./table";
 import TreeTable from "./tree-table";
 
 export {
